Move session status from footer to header
